FWMT-430 Refactored CSVParsingService
