+++ build.gradle
@@ -134,14 +134,17 @@ processResources {
 jacocoTestReport {
     reports {
         xml.enabled true
-        html.enabled false
+        html.enabled true
     }
     afterEvaluate {
         classDirectories = files(classDirectories.files.collect {
             fileTree(dir: it, exclude: [
                     '**/legacy_ingest/*',
                     '**/dto/*',
-                    '**/data/*',
+                    '**/data/tm/*',
+                    '**/config/*',
+  					'**/ApplicationConfig.java',
+  					'**/WebConfig.java'
             ])
         })
     }
