--- src/main/java/uk/gov/ons/fwmt/job_service/data/csv_parser/LegacySampleIterator.java
@@ -1,169 +1,0 @@
-package uk.gov.ons.fwmt.job_service.data.csv_parser;
-
-import java.lang.annotation.Annotation;
-import java.lang.reflect.Field;
-import java.time.LocalDate;
-import java.util.Arrays;
-import java.util.Optional;
-
-import org.apache.commons.csv.CSVParser;
-import org.apache.commons.csv.CSVRecord;
-import org.springframework.beans.PropertyAccessor;
-import org.springframework.beans.PropertyAccessorFactory;
-
-import uk.gov.ons.fwmt.job_service.data.annotation.CSVColumn;
-import uk.gov.ons.fwmt.job_service.data.legacy_ingest.LegacySampleGFFDataIngest;
-import uk.gov.ons.fwmt.job_service.data.legacy_ingest.LegacySampleIngest;
-import uk.gov.ons.fwmt.job_service.data.legacy_ingest.LegacySampleLFSDataIngest;
-import uk.gov.ons.fwmt.job_service.data.legacy_ingest.LegacySampleSurveyType;
-import uk.gov.ons.fwmt.job_service.exceptions.types.FWMTCommonException;
-import uk.gov.ons.fwmt.job_service.rest.client.dto.FieldPeriodDto;
-
-public class LegacySampleIterator<LegacySampleIngest> extends CSVIterator<LegacySampleIngest> {
-
-  private LegacySampleSurveyType legacySampleSurveyType;
-  private LegacySampleIngest instance;
-
-  protected LegacySampleIterator(CSVParser parser, LegacySampleSurveyType legacySampleSurveyType) {
-    super(parser);
-    this.legacySampleSurveyType = legacySampleSurveyType;
-    instance = new LegacySampleIngest();
-  }
-
-  @Override
-  public LegacySampleIngest ingest(CSVRecord record) throws FWMTCommonException {
-    // handle fields specific to a survey type
-    switch (legacySampleSurveyType) {
-    case LFS:
-      parseLegacySampleLFSData(instance, record);
-      break;
-    case GFF:
-      parseLegacySampleGFFData(instance, record);
-      break;
-    default:
-      throw new IllegalArgumentException("Unknown survey type");
-    }
-    // derive the TM job id
-    instance.setTmJobId(constructTmJobId(record, legacySampleSurveyType));
-    // derive the coordinates, if we were given a non-null non-empty grid ref
-    if (instance.getOsGridRef() != null && instance.getOsGridRef().length() > 0) {
-      String[] osGridRefSplit = instance.getOsGridRef().split(",", 2);
-      if (osGridRefSplit.length != 2) {
-        throw FWMTCommonException
-            .makeCsvInvalidFieldException("OSGridRef", "Did not match the expected format of 'X,Y'");
-      }
-      instance.setGeoX(Float.parseFloat(osGridRefSplit[0]));
-      instance.setGeoY(Float.parseFloat(osGridRefSplit[1]));
-    }
-    return instance;
-  }
-  
-  
-  private void parseLegacySampleGFFData(LegacySampleIngest instance, CSVRecord record) throws FWMTCommonException {
-    // set normal fields
-    setFromCSVColumnAnnotations(instance, record, "GFF");
-    // set derived due date
-    LocalDate date = convertToFieldPeriodDate(instance.getStage());
-    instance.setDueDate(date);
-    instance.setCalculatedDueDate(String.valueOf(date));
-    // set survey type and extra data
-    instance.setLegacySampleSurveyType(LegacySampleSurveyType.GFF);
-    instance.setGffData(new LegacySampleGFFDataIngest());
-    instance.setLfsData(null);
-    setFromCSVColumnAnnotations(instance.getGffData(), record, null);
-  }
-
-  private void parseLegacySampleLFSData(LegacySampleIngest instance, CSVRecord record) throws FWMTCommonException {
-    // set normal fields
-    setFromCSVColumnAnnotations(instance, record, "LFS");
-    // set derived due date
-    LocalDate date = convertToFieldPeriodDate(instance.getStage());
-    instance.setDueDate(date);
-    instance.setCalculatedDueDate(String.valueOf(date));
-    // set survey type and extra data
-    instance.setLegacySampleSurveyType(LegacySampleSurveyType.LFS);
-    instance.setGffData(null);
-    instance.setLfsData(new LegacySampleLFSDataIngest());
-    setFromCSVColumnAnnotations(instance.getLfsData(), record, null);
-  }
-
-  /**
-   * Create a unique Job ID that can be used by TotalMobile from existing fields
-   * within the CSV The method varies on the type of survey
-   */
-  protected String constructTmJobId(CSVRecord record, LegacySampleSurveyType surveyType) {
-    switch (surveyType) {
-    case GFF:
-      // quota '-' addr '-' stage
-      return record.get("TLA") + "-" + record.get("Quota") + "-" + record.get("AddressNo") + "-" + record.get("Stage");
-    case LFS:
-      // quota week w1yr qrtr addr wavfnd hhld chklet
-      if (!record.get("Issue_No").equals("1")) {
-        return record.get("QUOTA") + " " + record.get("WEEK") + " " + record.get("W1YR") + " " + record.get("QRTR")
-            + " "
-            + record.get("ADDR")
-            + " " + record.get("WAVFND") + " " + record.get("HHLD") + " " + record.get("CHKLET") + " - "
-            + record.get("FP") + " [R" + record.get("Issue_No") + "]";
-      } else {
-        return record.get("QUOTA") + " " + record.get("WEEK") + " " + record.get("W1YR") + " " + record.get("QRTR")
-            + " "
-            + record.get("ADDR")
-            + " " + record.get("WAVFND") + " " + record.get("HHLD") + " " + record.get("CHKLET") + " - "
-            + record.get("FP");
-      }
-    default:
-      throw new IllegalArgumentException("Invalid survey type");
-    }
-  }
-
-  public LocalDate convertToFieldPeriodDate(String stage) throws FWMTCommonException{
-    final Optional<FieldPeriodDto> existsByFieldperiod = fieldPeriodResourceServiceClient.findByFieldPeriod(stage);
-    if (existsByFieldperiod.isPresent()) {
-      final FieldPeriodDto fieldPeriod = existsByFieldperiod.get();
-      return fieldPeriod.getEndDate();
-    } else {
-      throw FWMTCommonException.makeUnknownFieldPeriodException(stage);
-    }
-  }
-  protected <T> void setFromCSVColumnAnnotations(T instance, CSVRecord record, String pivot)
-      throws FWMTCommonException {
-    Class<?> tClass = instance.getClass();
-    PropertyAccessor accessor = PropertyAccessorFactory.forBeanPropertyAccess(instance);
-    for (Field field : tClass.getDeclaredFields()) {
-      Optional<Annotation> annotation = Arrays.stream(field.getDeclaredAnnotations())
-          .filter(an -> an.annotationType() == CSVColumn.class)
-          .findFirst();
-      if (annotation.isPresent()) {
-        CSVColumn csvColumn = (CSVColumn) annotation.get();
-        String columnName;
-        if (!csvColumn.value().isEmpty()) {
-          columnName = csvColumn.value();
-        } else if (csvColumn.values().length > 0) {
-          Optional<CSVColumn.Mapping> mapping = Arrays.stream(csvColumn.values())
-              .filter(m -> m.when().equals(pivot))
-              .findFirst();
-          if (mapping.isPresent()) {
-            columnName = mapping.get().value();
-          } else {
-            throw FWMTCommonException.makeCsvOtherException("Given pivot does not occur in the CSVColumn annotation");
-          }
-        } else {
-          throw FWMTCommonException.makeCsvOtherException("CSVColumn lacked a 'value' or 'values' field");
-        }
-        // if it's mandatory or set, try
-        // if it's ignored, don't try
-        if (!csvColumn.ignored()) {
-          if (record.isSet(columnName)) {
-            accessor.setPropertyValue(field.getName(), record.get(columnName));
-          } else {
-            if (csvColumn.mandatory()) {
-              throw FWMTCommonException.makeCsvMissingColumnException(columnName);
-            }
-          }
-        }
-      }
-    }
-  }
-
-  
-}
