+++ src/main/java/uk/gov/ons/fwmt/job_service/service/impl/JobProcessor.java
@@ -1,24 +1,7 @@
 package uk.gov.ons.fwmt.job_service.service.impl;
 
-import com.consiliumtechnologies.schemas.services.mobile._2009._03.messaging.SendCreateJobRequestMessage;
-import com.consiliumtechnologies.schemas.services.mobile._2009._03.messaging.SendUpdateJobHeaderRequestMessage;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.scheduling.annotation.Async;
-import org.springframework.stereotype.Component;
-import uk.gov.ons.fwmt.job_service.data.csv_parser.CSVParseResult;
-import uk.gov.ons.fwmt.job_service.data.file_ingest.SampleFilenameComponents;
-import uk.gov.ons.fwmt.job_service.data.legacy_ingest.LegacySampleIngest;
-import uk.gov.ons.fwmt.job_service.exceptions.ExceptionCode;
-import uk.gov.ons.fwmt.job_service.exceptions.types.FWMTCommonException;
-import uk.gov.ons.fwmt.job_service.rest.client.JobResourceServiceClient;
-import uk.gov.ons.fwmt.job_service.rest.client.UserResourceServiceClient;
-import uk.gov.ons.fwmt.job_service.rest.client.dto.JobDto;
-import uk.gov.ons.fwmt.job_service.rest.client.dto.UserDto;
-import uk.gov.ons.fwmt.job_service.service.CSVParsingService;
-import uk.gov.ons.fwmt.job_service.service.totalmobile.TMJobConverterService;
-import uk.gov.ons.fwmt.job_service.service.totalmobile.TMService;
-import uk.gov.ons.fwmt.job_service.utils.SampleFileUtils;
+import static uk.gov.ons.fwmt.job_service.exceptions.types.FWMTCommonException.JOB_ENTRY_FAILED_STRING;
+import static uk.gov.ons.fwmt.job_service.exceptions.types.FWMTCommonException.JOB_FAILED_STRING;
 
 import java.io.File;
 import java.io.FileInputStream;
@@ -31,15 +14,32 @@ import java.time.format.DateTimeFormatter;
 import java.util.Iterator;
 import java.util.Optional;
 
-import static uk.gov.ons.fwmt.job_service.exceptions.types.FWMTCommonException.JOB_ENTRY_FAILED_STRING;
-import static uk.gov.ons.fwmt.job_service.exceptions.types.FWMTCommonException.JOB_FAILED_STRING;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.scheduling.annotation.Async;
+import org.springframework.stereotype.Component;
+
+import com.consiliumtechnologies.schemas.services.mobile._2009._03.messaging.SendCreateJobRequestMessage;
+import com.consiliumtechnologies.schemas.services.mobile._2009._03.messaging.SendUpdateJobHeaderRequestMessage;
+
+import lombok.extern.slf4j.Slf4j;
+import uk.gov.ons.fwmt.job_service.data.csv_parser.CSVParseResult;
+import uk.gov.ons.fwmt.job_service.data.csv_parser.CSVParserBuilder;
+import uk.gov.ons.fwmt.job_service.data.file_ingest.SampleFilenameComponents;
+import uk.gov.ons.fwmt.job_service.data.legacy_ingest.LegacySampleIngest;
+import uk.gov.ons.fwmt.job_service.exceptions.ExceptionCode;
+import uk.gov.ons.fwmt.job_service.exceptions.types.FWMTCommonException;
+import uk.gov.ons.fwmt.job_service.rest.client.FieldPeriodResourceServiceClient;
+import uk.gov.ons.fwmt.job_service.rest.client.JobResourceServiceClient;
+import uk.gov.ons.fwmt.job_service.rest.client.UserResourceServiceClient;
+import uk.gov.ons.fwmt.job_service.rest.client.dto.JobDto;
+import uk.gov.ons.fwmt.job_service.rest.client.dto.UserDto;
+import uk.gov.ons.fwmt.job_service.service.totalmobile.TMJobConverterService;
+import uk.gov.ons.fwmt.job_service.service.totalmobile.TMService;
+import uk.gov.ons.fwmt.job_service.utils.SampleFileUtils;
 
 @Slf4j
 @Component
 public class JobProcessor {
-
-  @Autowired
-  private CSVParsingService csvParsingService;
 
   @Autowired
   private UserResourceServiceClient userResourceServiceClient;
@@ -52,13 +52,16 @@ public class JobProcessor {
 
   @Autowired
   private TMService tmService;
+  
+  @Autowired
+  private FieldPeriodResourceServiceClient fieldPeriodResourceServiceClient;
 
   @Async("processExecutor")
   public void processSampleFile(File file) throws IOException {
     SampleFilenameComponents filename = SampleFileUtils.buildSampleFilenameComponents(file);
     final Reader reader = new InputStreamReader(new FileInputStream(file), StandardCharsets.UTF_8);
 
-    Iterator<CSVParseResult<LegacySampleIngest>> csvRowIterator = csvParsingService.parseLegacySample(reader, filename.getTla());
+    Iterator<CSVParseResult<LegacySampleIngest>> csvRowIterator = CSVParserBuilder.buildLegacySampleParserIterator(reader, filename.getTla(), fieldPeriodResourceServiceClient);
 
     while (csvRowIterator.hasNext()) {
       CSVParseResult<LegacySampleIngest> row = csvRowIterator.next();
