+++ src/test/java/uk/gov/ons/fwmt/job_service/data/csv_parser/legacysample/LegacySampleIteratorTests.java
@@ -1,0 +1,5 @@
+package uk.gov.ons.fwmt.job_service.data.csv_parser.legacysample;
+
+public class LegacySampleIteratorTests {
+
+}
